convert `magic8ball` to slash command
